# Optimizations to P256 operations