<!-- Thank you for your interest in contributing to OpenZeppelin! -->

<!-- Consider opening an issue for discussion prior to submitting a PR. -->
<!-- New features will be merged faster if they were first discussed and designed with the team. -->

Optimizes some `P256` operations by removing single-used variables and instead using the space for values that are used multiple times. 

### Gas report

#### Before
| test/utils/cryptography/P256.t.sol:GasMeasurer contract |                 |        |        |        |         |
|---------------------------------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                                         | Deployment Size |        |        |        |         |
| 638441                                                  | 2736            |        |        |        |         |
| Function Name                                           | min             | avg    | median | max    | # calls |
| verify                                                  | 255695          | 260876 | 261075 | 265421 | 256     |
| verifySolidity                                          | 251722          | 256903 | 257102 | 261448 | 256     |

#### After
| test/utils/cryptography/P256.t.sol:GasMeasurer contract |                 |        |        |        |         |
|---------------------------------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                                         | Deployment Size |        |        |        |         |
| 631952                                                  | 2706            |        |        |        |         |
| Function Name                                           | min             | avg    | median | max    | # calls |
| verify                                                  | 250795          | 256630 | 256547 | 261111 | 256     |
| verifySolidity                                          | 246822          | 252657 | 252574 | 257138 | 256     |

#### Difference

- `verify`: average gas reduced by 1.64% (4246 gas)
- `verifySolidity`: average gas reduced by 1.67% (4246 gas)
- Deployment cost: reduced by 1.02% (6489 gas)

<details>
  <summary>Replicate</summary>
Modify the P256 foundry tests slightly to perform the operations via a contract that uses the library (in order to be able to generate the gas report with forge):

```solidity
contract GasMeasurer {
    function verify(bytes32 digest, bytes32 r, bytes32 s, bytes32 x, bytes32 y) external returns (bool) {
        return P256.verify(digest, r, s, x, y);
    }

    function verifySolidity(bytes32 digest, bytes32 r, bytes32 s, bytes32 x, bytes32 y) external returns (bool) {
        return P256.verifySolidity(digest, r, s, x, y);
    }
}

contract P256Test is Test {
    /// forge-config: default.fuzz.runs = 512
    function testVerify(uint256 seed, bytes32 digest) public {
        GasMeasurer gas = new GasMeasurer();
        uint256 privateKey = bound(uint256(keccak256(abi.encode(seed))), 1, P256.N - 1);

        (uint256 x, uint256 y) = vm.publicKeyP256(privateKey);
        (bytes32 r, bytes32 s) = vm.signP256(privateKey, digest);
        s = _ensureLowerS(s);
        assertTrue(gas.verify(digest, r, s, bytes32(x), bytes32(y)));
        assertTrue(gas.verifySolidity(digest, r, s, bytes32(x), bytes32(y)));
    }
```

Then run: `forge t --mc P256Test --gas-report`
</details>

<!-- Describe the changes introduced in this pull request. -->
<!-- Include any context necessary for understanding the PR's purpose. -->


#### PR Checklist

<!-- Before merging the pull request all of the following must be complete. -->
<!-- Feel free to submit a PR or Draft PR even if some items are pending. -->
<!-- Some of the items may not apply. -->

- [ ] Tests
- [ ] Documentation
- [ ] Changeset entry (run `npx changeset add`)